requirements.txt
streamlit>=1.35.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.127.0
pandas>=2.2.0
